Dump configuration, options cleanup, sell/refine%.
The CPP DB Dump can now be configured using the `-ccpDump` flag which
will be used when the `--forceDB` flag is given or the `eve.h2.db` file
isn't around.

The `--asteroid` flag was renamed to `--material` making it more
generic. However the database hasn't been updated with any materials
other than asteroids and ores.

The better to buy or better to refine message now states the percentage
by how much better.

TODO
---------------------------------------------------------------------
 * Prices are hard-coded to Verge Vendor currently.
 * Station standing is hard-coded at 1.2.
 * Corp tax rate is hard-coded at 11%.
 * The printing to the table qualifies for a WTF award.
 * I'd like to show the station where the prices were quoted at.
 * I'd like 50% automatically shown, this is useful for teams.
 * There should be a verbose mode and a quite mode.
 * Need to figure out where EVE does it's rounding. Doing too much.